added lots of comments and moved files around
