+++ src/lib/dubins/tests/dubinsWaypoint.test.ts
@@ -1,7 +1,7 @@
 import { expect, test } from "bun:test";
-import { applyBounds, dubinsBetweenDubins, getBounds, getMinTurnRadius, getTunableDubinsParameters, localisePath, splitDubinsRuns } from "./dubinWaypoints";
+import { applyBounds, dubinsBetweenDubins, getBounds, getMinTurnRadius, getTunableDubinsParameters, localisePath, splitDubinsRuns } from "@/lib/dubins/dubinWaypoints";
 import { Waypoint } from "@/types/waypoints";
-import { defaultWaypoint } from "../waypoints/defaults";
+import { defaultWaypoint } from "@/lib/waypoints/defaults";
 import { dubinsPoint, Path, XY } from "@/types/dubins";
 import { defaultPlane } from "@/util/defaultVehicles";
