+++ src/lib/optimisation/gradient.ts
