+++ src/lib/waypoints/fns.ts
