+++ src/lib/world/conversion.ts
@@ -1,6 +1,11 @@
 import { LatLng, XY } from "@/types/dubins";
 
-// Convert WGS84 to local ENU using a given reference point
+/**
+ * Convert WGS84 to local ENU using a given reference point
+ * @param {LatLng} reference - The reference point in WGS84 coordinates (latitude/longitude)
+ * @param {LatLng} point - The point in WGS84 coordinates (latitude/longitude)
+ * @returns {XY} The converted local ENU coordinates
+ */
 export function g2l(reference: LatLng, point: LatLng): XY {
   const { lat: refLat, lng: refLng } = reference;
   const { lat, lng } = point;
@@ -19,7 +24,12 @@ export function g2l(reference: LatLng, point: LatLng): XY {
   return { y: north, x: east };
 }
 
-// Convert local ENU to WGS84 using a given reference point
+/**
+ * Convert local ENU to WGS84 using a given reference point
+ * @param {LatLng} reference - The reference point in WGS84 coordinates (latitude/longitude)
+ * @param {XY} point - The point in local ENU coordinates (x=east, y=north in meters)
+ * @returns {LatLng} The converted WGS84 coordinates
+ */
 export function l2g(reference: LatLng, point: XY): LatLng {
   const { lat: refLat, lng: refLng } = reference;
   const { x: east, y: north } = point;
